Update project name and dependencies, delete .env.example file
The project name in package.json and package-lock.json has been updated to "collateralized-loan-project". Version was also added to the project and Ether's dependency version has been updated in the same files. A new script for testing was added to package.json. In addition, the .env.example file was removed.